requirement.txt
jittor>=1.3.4.4
numpy
tqdm
opencv-python
Pillow
imageio
pyyaml